Fixed writable type.

+++ src/UnoDriver.java
@@ -1,6 +1,7 @@
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -35,7 +36,7 @@ public class UnoDriver extends Configured implements Tool {
 		job.setMapOutputValueClass(DoubleWritable.class);
 
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(DoubleWritable.class);
+		job.setOutputValueClass(IntWritable.class);
 
 		boolean success = job.waitForCompletion(true);
 		return success ? 0 : 1;
